virtual machine prototype
refactor wolfBoot_start to operate as a virtual machine that steps through
states that can be saved and repeated if interrupted. a saved state is
redundant and includes a digest to verify it wasn't corrupted during write.
using states like this means it no longer needs to use sector status which
may become corrupted during powerfailure. it also allows us to verify any
has completed an not just sector writes so auxilary actions like saving keys
are gaurenteed to complete